ref: move nodestore.bigtable.backend test so it actually runs
